splitting: new module for WordSplitter and implementations
